requirements.txt
streamlit
numpy
pillow
matplotlib
pandas
scikit-learn